requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-filter>=24.2
dj-database-url>=2.2
gunicorn>=21.2
psycopg2-binary>=2.9
python-dotenv>=1.0
whitenoise>=6.8
openpyxl>=3.1
pypdf>=5.0
